Bug 1275064 - Use unsigned type in RealWord bitfield. r=froydnj
UltraBlame original commit: 0928a16792af6f0a947b1fdca5b12d16b84eaca0

+++ extensions/spellcheck/src/mozInlineSpellWordUtil.cpp
@@ -783,7 +783,7 @@ mozInlineSpellWordUtil::FindRealWordContaining(int32_t aSoftTextOffset,
   
   const RealWord& word = mRealWords[index];
   int32_t offset = aSoftTextOffset - word.mSoftTextOffset;
-  if (offset >= 0 && offset <= word.mLength)
+  if (offset >= 0 && offset <= static_cast<int32_t>(word.mLength))
     return index;
 
   if (aSearchForward) {
